tests/server: fix `-Wconversion`
The hope is that this fixes the FD_SET() type misalignment, that
is present in MSYS but not in Cygwin builds:

```
../../../docs/examples/sendrecv.c:47:3: warning: conversion to 'long unsigned int' from 'curl_socket_t' {aka 'int'} may change the sign of the result [-Wsign-conversion]
47 |   FD_SET(sockfd, &errfd); /* always check for error */
   |   ^~~~~~
```
Ref: https://ci.appveyor.com/project/curlorg/curl/builds/48794906/job/54616tw3d74xrf6o?fullLog=true#L3031

This has been fixed in Cygwin in mid-2020:
https://cygwin.com/git/?p=newlib-cygwin.git;a=commitdiff;h=5717262b8ecfed0f7fab63e2c09c78991e36f9dd

+++ tests/server/mqttd.c
@@ -497,7 +497,7 @@ static curl_socket_t mqttit(curl_socket_t fd)
   unsigned short packet_id;
   size_t payload_len;
   size_t client_id_length;
-  unsigned int topic_len;
+  size_t topic_len;
   size_t remaining_length = 0;
   size_t bytes = 0; /* remaining length field size in bytes */
   char client_id[MAX_CLIENT_ID_LENGTH];
@@ -635,7 +635,7 @@ static curl_socket_t mqttit(curl_socket_t fd)
       /* two bytes topic length */
       topic_len = (size_t)(buffer[2] << 8) | buffer[3];
       if(topic_len != (remaining_length - 5)) {
-        logmsg("Wrong topic length, got %u expected %zu",
+        logmsg("Wrong topic length, got %zu expected %zu",
                topic_len, remaining_length - 5);
         goto end;
       }
